agents/requirements.txt
azure-ai-projects==1.0.0b11
azure-identity==1.23.0
aiohttp==3.12.7
pydantic==2.11.5
python-dotenv==1.0.1
azure-cosmos==4.9.0